duration-2 refactor current test
