+++ src/test/java/org/mikeneck/duration/AppTest.java
@@ -3,7 +3,6 @@
  */
 package org.mikeneck.duration;
 
-import java.io.ByteArrayOutputStream;
 import java.nio.charset.StandardCharsets;
 import java.time.Clock;
 import java.time.Duration;
@@ -11,6 +10,8 @@ import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mikeneck.duration.util.StdOut;
+import org.mikeneck.duration.util.StdOutReplacer;
 import picocli.CommandLine;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -19,7 +20,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 class AppTest {
 
     @Test
-    void normal(ByteArrayOutputStream out) {
+    void normal(StdOut out) {
         OffsetDateTime input = OffsetDateTime.of(2020, 5, 4, 15, 2, 1, 0, ZoneOffset.UTC);
         Duration duration = Duration.parse("PT-128H-45M-30S");
         OffsetDateTime now = input.minus(duration);
@@ -27,7 +28,7 @@ class AppTest {
         App app = new App(clock);
         int result = new CommandLine(app).execute("2020-05-04T15:02:01Z");
         assertEquals(0, result);
-        String stdOut = out.toString(StandardCharsets.UTF_8);
+        String stdOut = out.get().toString(StandardCharsets.UTF_8);
         assertEquals("PT-128H-45M-30S\n", stdOut);
     }
 }
