+++ src/test/java/org/mikeneck/duration/util/StdOut.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2020 Shinya Mochida
+ *
+ * Licensed under the Apache License,Version2.0(the"License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,software
+ * Distributed under the License is distributed on an"AS IS"BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.mikeneck.duration.util;
+
+import java.io.ByteArrayOutputStream;
+import org.jetbrains.annotations.NotNull;
+
+@FunctionalInterface
+public interface StdOut {
+
+    @NotNull ByteArrayOutputStream get();
+}
